JAMES-3491 Transport independant JMAP API
Allows reuse over HTTP or WebSocket transport
